added player and some other minor changes
so far we can now switch between rooms by walking up to the door of a house and hitting return key.
i have created two houses which lead to the same room due to lack of another room(anyone can create rooms in the rooms scene at a different area on the scene).