Adds the Interactive Button behaviour and applies to Menu and Settings (#54)
Imported a new Extension called Button, which has a Interactive Button
behaviour for making Sprites into regular GUI buttons.
This new behaviour encapsulates handling of mouse up, down, hover and
button disabled states. The behaviour can be configured to set which
animations (by name) will be used for each button state, simplifying
the events sheet.

The behaviour was applied to the buttons of Settings and Menu scenes.
However, issues in these Scenes are not fixed by this commit.